chromium/third_party/blink/web_tests/webcodecs/basic_video_encoding.html

<!DOCTYPE html>
<html>
<head>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
</head>
<body>
<img id='frame_image' style='display: none;' src="pattern.png">
<script>

'use strict';

async function generateBitmap(width, height, text) {
  let img = document.getElementById('frame_image');
  let cnv = document.createElement("canvas");
  cnv.height = height;
  cnv.width = width;
  var ctx = cnv.getContext('2d');
  ctx.drawImage(img, 0, 0, width, height);
  ctx.font = '30px fantasy';
  ctx.fillText(text, 5, 40);
  return createImageBitmap(cnv);
}

async function createFrame(width, height, ts) {
  let imageBitmap = await generateBitmap(width, height, ts.toString());
  return new VideoFrame(imageBitmap, { timestamp: ts });
}

async function encode_test(codec, acc) {
  let w = 640;
  let h = 480;
  let next_ts = 0
  let frames_to_encode = 20;
  let frames_processed = 0;
  let process_video_chunk = function(chunk) {
    assert_greater_than_equal(chunk.timestamp, next_ts++);
    let type = (chunk.timestamp % 5 == 0) ? "key" : "delta";
    assert_equals(chunk.type, type);
    var data = new Uint8Array(chunk.data);
    assert_greater_than_equal(data.length, 0);
    console.log("data len: " + data.length);
    frames_processed++;
  };

  const init = {
    output : process_video_chunk,
    error: (e) => { console.log(e.message); }
  };
  const params = {
    codec : codec,
    acceleration: acc,
    width : w,
    height : h,
    bitrate: 10e6,
    framerate: 30,
  };
  let encoder = new VideoEncoder(init);
  encoder.configure(params);
  for (let i = 0; i < frames_to_encode; i++) {
    var frame = await createFrame(w, h, i);
    let keyframe = (i % 5 == 0);
    encoder.encode(frame, { keyFrame : keyframe});
  }
  await encoder.flush();
  encoder.close();
  assert_equals(frames_processed, frames_to_encode);
}

promise_test(encode_test.bind(null, "vp09.00.10.08", "allow"),
    "encoding vp9 profile0");

promise_test(encode_test.bind(null, "vp09.02.10.10", "allow"),
    "encoding vp9 profile2");

promise_test(encode_test.bind(null, "vp8", "allow"),
    "encoding vp8");


/* Uncomment this for manual testing, before we have GPU tests for that
promise_test(encode_test.
  bind(null, "avc1.42001E", "avc1.42001E", "require"),
  "encoding avc1.42001E");
*/

</script>
</body>
</html>
